+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Timer Interrupts are now handled in TIFR0 and TIMSK0, too.
+
 ## v.0.4.2 - 2020-11-13
 
 - Added two functions to system_t: